requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv
psycopg2-binary
